+++ src/main/java/depth/jeonsilog/domain/exhibition/application/ExhibitionService.java
@@ -52,11 +52,7 @@ public class ExhibitionService {
     private final RatingRepository ratingRepository;
 
     private final UserService userService;
-    private final S3Uploader s3Uploader;
 
-    private static final String DIRNAME = "exhibition_img";
-
-    // Description : 전시회 목록 조회 - 요즘 뜨는 전시
     public ResponseEntity<?> findRecentlyExhibitionList(Integer page) {
 
         PageRequest pageRequest = PageRequest.of(page, 10, Sort.by(
@@ -85,7 +81,6 @@ public class ExhibitionService {
         return ResponseEntity.ok(apiResponse);
     }
 
-    // Description : 전시회 목록 조회 - 다채로운 예술의 향연
     public ResponseEntity<?> findColorfulExhibitionList() {
 
         List<Exhibition> exhibitions = exhibitionRepository.findExhibitionsByAddressContainingKeyword("서울특별시 종로구");
@@ -142,7 +137,6 @@ public class ExhibitionService {
         return ResponseEntity.ok(apiResponse);
     }
 
-    // Description : 전시회 목록 조회 - 곧 종료되는 전시
     public ResponseEntity<?> findEndingSoonExhibitionList() {
 
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyyMMdd");
@@ -188,7 +182,6 @@ public class ExhibitionService {
         return ResponseEntity.ok(apiResponse);
     }
 
-    // Description : 전시회 목록 조회 - 새로 시작한 전시
     public ResponseEntity<?> findNewExhibitionList() {
 
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyyMMdd");
@@ -234,7 +227,6 @@ public class ExhibitionService {
         return ResponseEntity.ok(apiResponse);
     }
 
-    // Description : 전시회 상세 정보 조회
     public ResponseEntity<?> findExhibition(UserPrincipal userPrincipal, Long exhibitionId) {
 
         User user = userService.validateUserByToken(userPrincipal);
@@ -260,7 +252,6 @@ public class ExhibitionService {
         return ResponseEntity.ok(apiResponse);
     }
 
-    // Description : 랜덤 전시회 2개 조회
     public ResponseEntity<?> randomTwoExhibitions() {
 
         long size = exhibitionRepository.count();
@@ -310,7 +301,6 @@ public class ExhibitionService {
         return ResponseEntity.ok(apiResponse);
     }
 
-    // Description : 검색어를 포함한 전시회 목록 조회
     public ResponseEntity<?> searchExhibitions(Integer page, String searchWord) {
 
         PageRequest pageRequest = PageRequest.of(page, 10, Sort.by(Sort.Direction.ASC, "createdDate"));
@@ -338,7 +328,6 @@ public class ExhibitionService {
         return ResponseEntity.ok(apiResponse);
     }
 
-    // Description : 전시회 이름만으로 전시회 목록 조회 - 관리자 페이지, 포토캘린더에서 사용한다.
     public ResponseEntity<?> searchExhibitionsByName(Integer page, String exhibitionName) {
 
         PageRequest pageRequest = PageRequest.of(page, 10);
@@ -352,73 +341,6 @@ public class ExhibitionService {
         ExhibitionResponseDto.SearchExhibitionByNameResListWithPaging searchExhibitionByNameResListWithPaging = ExhibitionConverter.toSearchExhibitionByNameResListWithPaging(hasNextPage, exhibitionResList);
 
         ApiResponse apiResponse = ApiResponse.toApiResponse(searchExhibitionByNameResListWithPaging);
-        return ResponseEntity.ok(apiResponse);
-    }
-
-    // Description : 전시회 상세 정보 수정
-    @Transactional
-    public ResponseEntity<?> updateExhibitionDetail(UserPrincipal userPrincipal, ExhibitionRequestDto.UpdateExhibitionDetailReq updateExhibitionDetailReq, MultipartFile img) throws IOException {
-
-        User user = userService.validateUserByToken(userPrincipal);
-        DefaultAssert.isTrue(user.getRole().equals(Role.ADMIN), "관리자만 수정할 수 있습니다.");
-
-        Exhibition exhibition = validateExhibitionById(updateExhibitionDetailReq.getExhibitionId());
-
-        String storedFileName = null;
-
-        if (updateExhibitionDetailReq.getIsImageChange()) { // 이미지를 변경하는 경우
-            if (img != null) { // 이미지 삭제가 아닌 이미지를 변경하거나 추가하는 경우
-                storedFileName = s3Uploader.upload(img, DIRNAME);
-            }
-            // 기존 포스터 이미지가 s3에 있으면, 이미지 삭제 / 없으면(OPEN API 포스터 이미지 or NULL의 경우) 말고
-            s3Uploader.deleteImage(DIRNAME, exhibition.getImageUrl());
-        }
-
-        exhibition.updateExhibitionDetail(updateExhibitionDetailReq, storedFileName);
-
-        Place place = exhibition.getPlace();
-        place.updatePlaceWithExhibitionDetail(updateExhibitionDetailReq.getUpdatePlaceInfo());
-
-        ApiResponse apiResponse = ApiResponse.toApiResponse(
-                Message.builder().message("전시회 및 전시공간 정보를 수정했습니다.").build());
-
-        return ResponseEntity.ok(apiResponse);
-
-    }
-
-    // Description: 관리자 페이지 전시회 sequence 수정
-    @Transactional
-    public ResponseEntity<?> updateExhibitionSequence(UserPrincipal userPrincipal, ExhibitionRequestDto.UpdateExhibitionSequenceList updateSequenceReq) {
-
-        User findUser = userService.validateUserByToken(userPrincipal);
-        DefaultAssert.isTrue(findUser.getRole() == Role.ADMIN, "관리자만 전시회 순서를 변경할 수 있습니다.");
-
-        List<ExhibitionRequestDto.UpdateExhibitionSequence> updateExhibitionSequenceList = updateSequenceReq.getUpdateSequenceInfo();
-        int size = updateExhibitionSequenceList.size();
-        for (int i = 0; i < size; i++) {
-            ExhibitionRequestDto.UpdateExhibitionSequence updateExhibitionSequence = updateExhibitionSequenceList.get(i);
-            Optional<Exhibition> exhibition = exhibitionRepository.findById(updateExhibitionSequence.getExhibitionId());
-            DefaultAssert.isTrue(exhibition.isPresent(), "전시회 정보가 올바르지 않습니다.");
-
-            Optional<Exhibition> duplicatedExhibition = exhibitionRepository.findBySequence(updateExhibitionSequence.getSequence());
-            duplicatedExhibition.ifPresent(value -> value.updateSequence(11));  // 해당 sequence를 가지고 있던 기존의 전시회는 11번으로 순서를 변경한다.
-
-            Exhibition findExhibition = exhibition.get();
-            findExhibition.updateSequence(updateExhibitionSequence.getSequence());
-        }
-        // sequence 삭제 로직
-        for (int i = size; i < 10 ;i++) {
-            Optional<Exhibition> tempExhibition = exhibitionRepository.findBySequence(i + 1);
-            if (tempExhibition.isPresent()) {
-                tempExhibition.get().updateSequence(11);
-            } else {
-                break;
-            }
-        }
-
-        ApiResponse apiResponse = ApiResponse.toApiResponse(
-                Message.builder().message("전시회 순서를 변경했습니다.").build());
-
         return ResponseEntity.ok(apiResponse);
     }
 
@@ -454,5 +376,4 @@ public class ExhibitionService {
 
         return randomIndices;
     }
-
 }
