+++ src/main/java/depth/jeonsilog/domain/stop/presentation/StopController.java
@@ -1,13 +1,14 @@
 package depth.jeonsilog.domain.stop.presentation;
 
 import depth.jeonsilog.domain.stop.application.StopService;
-import depth.jeonsilog.domain.stop.dto.StopRequestDto;
 import depth.jeonsilog.global.config.security.token.CurrentUser;
 import depth.jeonsilog.global.config.security.token.UserPrincipal;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
@@ -15,14 +16,6 @@ import org.springframework.web.bind.annotation.*;
 public class StopController implements StopApi {
 
     private final StopService stopService;
-
-    @PostMapping
-    public ResponseEntity<?> stopUser(
-            @CurrentUser final UserPrincipal userPrincipal,
-            @Valid @RequestBody final StopRequestDto.StopUserReq dto
-    ) {
-        return stopService.stopUser(userPrincipal, dto);
-    }
 
     @GetMapping
     public ResponseEntity<?> findUserIsStop(
